Add creation of messages and group finders into concerns
It is now possible to create messages by sending a request to the path,
provided the receiver's uuid is given.

I have grouped the code for finding the receiver and the chat between them
and the current user withing two methods in the ChatFinders module to
avoid too much repetition and make the controller actions easier to read.