Add skeleton for Chat
I've added a receiver state on the parent that gets updated whenever a
user is clicked from the UserList. The Chat component takes this receiver
as a prop and looks for a chat between the receiver and the current user
whenever the receiver changes.

The returned jsx is, for now, a skeleton meant to give a foundation to build
upon.