Add ChatList component
It is structured in the same way as UserList, but with a different fetch
and different state name for clarity. I have decided to make two separate
components, in spite of how similar they are, because I plan to add
scroll loading soon and I did not want a single component to become too
bloated with different if clauses based on whether the list was being used
to render just Users or Users that have a chat with the current user.