Fix chat pagination bug
Fixed a bug that would cause the frontend to send a wrong pagination number
to the backend when a new receiver would be selected.

pagination.page would always increment by 1 on load and on the scroll event,
but would not reset when the user would click on a new receiver, so the
request for the new chat would be sent with, for example, page 5 instead
of page 1.