Add profile display for other users and fix chat bug
A user can now see the profile of another user either by hovering
on their card in the UserList on the right or by clicking on their name
after opening a chat with them.

I have also fixed the way chat is set after data retrieval; the previous
way was causing duplication if the user was returning to the chat after
visiting a profile.