Add mobile layout and fix eventListener bug
To add the mobile layout I've created a MobileLayout component that I am
conditionally rendering from the Home page depending on the width of the
window. The implementation is quite messy at the moment, with lots of
props being passed around and repeated, but as far as manual testing can
tell, it works functionally.

I've also finally figured out what was causing my eventListener removals
to missfire and it was the fact that the element they were attached to
were getting removed before the removal could fire, so I've now added a
nil check as a safeguard.

All the other changes are stylistic.