Fix ChatList not updating when a new chat was started
A refetchChatList state of the Home component is now changed by Chat when
a message is sent to a new chat. That state is then used by ChatList as
a useEffect condition that resets chatList pagination state, the ChatList
itself and fetches a new one to set the new chat on top of the list.