Group profile state and mobile chat display states
The profile display boolean and the user are now tracked and set by the
same unit. So is handling the chat display and the display of the chat list
on mobile.

Both of these felt like moves that belonged together, since the two states
were modified at the same time in the majority of cases and this has
reduced some of the state bloat.