Add search bar to UserList
I've added a search bar to the UserList component. It uses the real time
value of the input and sends a debounced request to the backend when the
user stops typing for 300ms.

I've had to update the #index method of my UsersController to search for
users based on the search param.

The next step will be to extract the search bar into a component and
add it to the chat list as well.