Add ActionCable for live chat updates
When the user opens a new chat, the getChat function will call the new
subscribeToChat function to subscribe the current user to that chat's room.

The chat room name is formed using said chat's id. To do this, I have had
to update the chats_controller to expose the id and turn the chat state
into an object with the keys "chat_id" and "messages".

This shouldn't be a security risk, because none of my routes use the id
of a chat to return data. Posting messages and getting chats is done
using current_user and the receiver's uuid.

The Connection class uses devise's warden to ensure that the user making
the request is the user that is currently logged in. The ChatChannel object
then checks that the user actually belongs to the chat that they are trying
to subscribe to.