Add ChatList updates when the user sends a message
This is all handled on the frontend via react state. I've had to lift the
chatList state into the home component and give it to both Chat and ChatList.

Now, when a user sends a message, the Chat component will filter the list
of chats to ensure that the current chat is not duplicated. Afterwards, it
sets chatList to a new array formed of the current chat and the filtered
version of the previous chat. This ensures that the chat the user has just
sent a message to appears at the top.

Before any of this, the function checks if the chat is already at the top
or if the user is on mobile, since in either of those cases the mutation
would be redundant.