Add notification icons next to unread messages
To achieve this, I have added a new #read column to my already existing
chat_memberships table. By default, this is set to true and changed to
false for the receiver when a message is sent.

A chat list refetch is triggered by my already existing WebSockets connection.
The data returned by that refetch will now also contain the read boolean.
If the latest messages have not been read, the frontend renders the icon.
The read status and icon display can update on two events: either when the
user clicks on the chat in the chat list, opening it, or when the user sends
a message in that chat.

Both of these serve their purpose: the first event should be obvious. If
the user has just clicked on a chat it means they are currently paying
attention to the app, they see the message and would like the icon to
go away.

The reason I still send notifications when the user has the chat open
and wait for a reply before removing the icon is in case the user alt tabs
or goes away for a while and might close the browser in a hurry when they
return. This ensures that the user doesn't miss a text due to being in
a rush or simply due to forgetting to reply. The icon will remain there
until they reply or deliberately click again on the chat in the chatList.