Add group chats
Users can now create group chats and talk in them. For scalability, I have
had to normalize the way data is transfered and used in my backend and
frontend. Receivers is now always an array, even if that array only
contains one member.

Sending messages, retrieving them in real time via web sockets and
notifying users when a message is sent in a group all work seamlessly,
same as in private chats.