Fix two bugs created by the latest changes
The first was that if the logged user was in the same group as another
user but didn't have a private chat with them, said user would not show
in their User List, since #index method of my #UsersController didn't
make a difference between private and group chats. It now does.

The second bug was a bug related to my #find_private_chat method in
 #ChatFinder. It turns out that my query for finding chats that specifically
 only have two users was wrong. In the scenario above, if the logged user
 clicked on the User List banner of a user they were in a chat with, but
 didn't have a private chat with, the group chat would be opened. The
 intended behavior was for that action to create a private chat, which
 updating the query has fixed.