Write searchBar tests and fix bugs
I've added tests for the implementation of the searchBar and fixed some
bugs found along the way:

In users_controller, #filter_users will now filter users not only by the
name of the users, but also by the name of their chat to ensure that they
can be found in the ChatList.

The relationship between ChatList/UserList and SearchBar has been modified.
The lists now have a searchTerm state that the SearchBar updates to trigger
the list refetch.
In my old implementation, because the lists didn't have a way to remember
the search term, only the first set of 20 elements would be filtered. As
the user scrolled down, page 2 of elements would not be filtered through
the search term.

The seeds have also been updated to facilitate testing.