Fix visual bugs with large group titles and userList on narrow screens
Large titles would cause the chat to overflow, so I've added overflow
handling with elipsis.

Since the overflow handling is present on every userHeader, this has
caused the UserList to become too small to contain its'header on narrow
screens, so I've also upped the flex growth of the container to have it
match the ChatList a bit more closely in width and prevent overflow
until the 700px treshold is reached.