Improve group functionality + resize handling for SearchBar
I've made it so the chat for a group is immediately opened after creation.
This has led to the discovery of several chained bugs that would prevent
a user from sending a message to a group without clicking on it in the
chatList first.
The cause for them was forgetting to update certain parts of code after
the data normalization I did a while back, but everything works fine now.

Users can create groups, the group chat will be opened automatically and
users will be able to send messages to them right away.

I've also added a way for UserList and ChatList that dynamically shows or
hides the searchBar based on the width of the window.