Add #show to events controller and event display flow for Profile
A user's events will now be displayed on their profile, split between
old and upcoming events.

Clicking on an event's banner will display a more detailed event page. A
user can backtrack to the profile by the click of a back button or, now
that the event's organisers are displayed, go to their profile by clicking
on their image.