Add pending_requests retrieval to users controller
The current_user_info method now returns pending requests to the frontend
and the user.events association will now only return events with a
"accepted" or "pending" status.

The data of accepted events will continue to be used for display, while
the data of pending events will be used for pending notifications.