Reworked EventForm to use the new UserListForm
EventForm will now make use of the SearchBar and UserListForm components
to allow the creator of an event to invite other organisers. The flow for
this is the same as the flow for creating a group: a display button renders
the form and dims the background and the form fetches a list of users that
can be staged or taken out of staging for sending invites.

The method for changing the event details is used as the callback prop for
this UserListForm.